+++ packages/core/src/public.ts
@@ -33,6 +33,8 @@ export type {
   SignTransactionResponse,
   SignAuthEntryOptions,
   SignAuthEntryResponse,
+  SignMessageOptions,
+  SignMessageResponse,
   SolanaSignInInput,
   SolanaSignInOutput,
   SolanaSignMessageResponse,